requirements.txt
PyOpenGL==3.1.7
pygame==2.5.2
numpy==1.26.4
pillow==11.3.0